requirements.txt
pip install requests
pip install python-dotenv
pip install Flask
pip install waitress

Package            Version
------------------ ---------
blinker            1.8.2
certifi            2024.8.30
charset-normalizer 3.4.0
click              8.1.7
colorama           0.4.6
Flask              3.0.3
idna               3.10
itsdangerous       2.2.0
Jinja2             3.1.4
MarkupSafe         3.0.2
pip                24.3.1
python-dotenv      1.0.1
requests           2.32.3
urllib3            2.2.3
waitress           3.0.1
Werkzeug           3.1.1

#creating and opening virtual environment
python -m venv .venv
Set-ExecutionPolicy -ExecutionPolicy RemoteSigned -Scope CurrentUser (If needed)
.venv\Scripts\activate
